backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.26.0
pillow==10.2.0
supabase==2.3.4
pydantic==2.5.3
pydantic-settings==2.1.0
